+++ app/src/main/java/QuestionBank/MultipleAnswerChoice.java
@@ -1,0 +1,15 @@
+package QuestionBank;
+
+import java.util.LinkedList;
+
+public class MultipleAnswerChoice extends Questions {
+    //Fields
+
+    //Constructor
+    protected MultipleAnswerChoice(int idNumber, String question, LinkedList<String> correctAnswers, LinkedList<String> possibleAnswers) {
+        super(idNumber, question, correctAnswers, possibleAnswers);
+    }
+
+    //Methods
+
+}