+++ Changelog.md
@@ -9,6 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 -
 
+## [0.1.0] - 10-11-2022
+### Added
+- Basic command line functionality
+- Ability to add a question to the bank
+- Ability to see all questions added to the bank
+
+### Changed
+- Nothing changed
+
+### Removed
+- Nothing removed
 
 ## [0.0.0] - 09-08-2022
 ### Added